@@ -12,6 +14,10 @@
 	<url>https://github.com/eduardovalentim/easymath</url>
 
 	<properties>
+		<!-- antlr4-runtime and antlr4-maven-plugin must match the same version 
+			to avoid errors at build time -->
+		<antlr4.version>4.10.1</antlr4.version>
+		<java.version>17</java.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 	</properties>
 
