@@ -2,8 +2,8 @@ package com.github.eduardovalentim.easymath.test;
 
 import java.math.BigDecimal;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class AlgebraMathTest {
 
@@ -14,7 +14,7 @@ public class AlgebraMathTest {
 		double expected = algebra.squaresDifference(7.3d, 3.7d);
 		BigDecimal actual = algebra.squaresDifferenceExpansion(new BigDecimal("7.3"), new BigDecimal("3.7"));
 
-		Assert.assertEquals(expected, actual.doubleValue(), 0.001);
+		Assertions.assertEquals(expected, actual.doubleValue(), 0.001);
 	}
 
 	@Test
@@ -24,7 +24,7 @@ public class AlgebraMathTest {
 		double expected = algebra.cubesDifference(7.3d, 3.7d);
 		BigDecimal actual = algebra.cubesDifferenceExpansion(new BigDecimal("7.3"), new BigDecimal("3.7"));
 
-		Assert.assertEquals(expected, actual.doubleValue(), 0.001);
+		Assertions.assertEquals(expected, actual.doubleValue(), 0.001);
 	}
 
 	@Test 
@@ -34,7 +34,7 @@ public class AlgebraMathTest {
 		double expected = algebra.cubesSum(7.3d, 3.7d);
 		BigDecimal actual = algebra.cubesSumExpansion(new BigDecimal("7.3"), new BigDecimal("3.7"));
 
-		Assert.assertEquals(expected, actual.doubleValue(), 0.001);
+		Assertions.assertEquals(expected, actual.doubleValue(), 0.001);
 	}
 
 }
